+++ kite/src/main/kotlin/com/tang/kite/config/defaults/DefaultTenantProcessor.kt
@@ -1,0 +1,17 @@
+package com.tang.kite.config.defaults
+
+import com.tang.kite.config.tenant.TenantProcessor
+import java.lang.reflect.Field
+
+/**
+ * Default tenant processor implementation
+ *
+ * @author Tang
+ */
+object DefaultTenantProcessor : TenantProcessor {
+
+    override fun getTenantIds(field: Field): List<Any> {
+        throw NotImplementedError("Please implement this method according to your business logic")
+    }
+
+}